Fix race with opertion cancellation.
Cancellation of index operations posts an ast in
idx_op_cb_cancel()->m0__idx_cancel(). There is no any synchronization with this
ast execution and as a result, m0_op_fini() might finalise the operation while
the ast is still in the fork queue. This leads to crashes in locality
m0_sm_asts_run().

The only thing the ast is doing is cancellation of the rpc items. There is no
good reason to do this in an ast, this can be done synchronously in
m0__idx_cancel().

Also, add a few checks that ast-s are completed when operations are finalised.

Signed-off-by: Nikita Danilov <nikita.danilov@seagate.com>